DeepSFNS/requirements.txt
absl-py==2.3.1
audioread==3.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
clarabel==0.11.1
contourpy==1.3.2
cvxpy==1.7.3
cycler==0.12.1
decorator==5.2.1
et_xmlfile==2.0.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
grpcio==1.75.1
h5py==3.15.0
idna==3.11
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lap==0.5.12
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
msgpack==1.1.2
munkres==1.1.4
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
openpyxl==3.1.5
osqp==1.0.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
protobuf==6.32.1
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
scs==3.2.9
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.5.1+cu118
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.9.2
webrtcvad==2.0.10
Werkzeug==3.1.3
